@@ -21,40 +21,40 @@
         "php": "^7.4|^8.0",
         "astrotomic/laravel-translatable": "^11.9.1",
         "box/spout": "^v3.3.0",
-        "doctrine/dbal": "^2.10",
+        "doctrine/dbal": "^2.13.9",
         "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^2.0.4",
-        "guzzlehttp/guzzle": "^7.3",
-        "intervention/image": "^2.6.1",
-        "laravel/framework": "^8.57",
-        "laravel/passport": "^10.1.3",
-        "laravel/tinker": "^2.6.1",
-        "laravel/ui": "^3.3",
-        "league/flysystem-aws-s3-v3": "^1.0.29",
-        "maxakawizard/json-collection-parser": "^1.8.0",
-        "mcamara/laravel-localization": "^1.6.1",
-        "meyfa/php-svg": "^0.11.2",
+        "fruitcake/laravel-cors": "^v2.2.0",
+        "guzzlehttp/guzzle": "^7.4.5",
+        "intervention/image": "^2.7.2",
+        "laravel/framework": "^v8.83.22",
+        "laravel/passport": "^v10.4.1",
+        "laravel/tinker": "^v2.7.2",
+        "laravel/ui": "^v3.4.6",
+        "league/flysystem-aws-s3-v3": "^1.0.30",
+        "maxakawizard/json-collection-parser": "^1.9.0",
+        "mcamara/laravel-localization": "^v1.7.0",
+        "meyfa/php-svg": "^v0.11.3",
         "myclabs/php-enum": "^1.8.3",
         "nikazooz/laravel-captcha": "^1.3",
-        "php-coord/php-coord": "^4.4",
-        "runalyze/dem-reader": "^1.1",
-        "spatie/laravel-activitylog": "^3.17",
-        "spatie/laravel-backup": "^6.16.2",
+        "php-coord/php-coord": "^v4.7.0",
+        "runalyze/dem-reader": "^v1.1.0",
+        "spatie/laravel-activitylog": "^3.17.0",
+        "spatie/laravel-backup": "^6.16.5",
         "spatie/laravel-cookie-consent": "^2.12.13",
         "spatie/laravel-menu": "^3.7.1",
-        "stevebauman/purify": "^4.0",
-        "tightenco/ziggy": "^0.9"
+        "stevebauman/purify": "^v4.0.1",
+        "tightenco/ziggy": "^0.9.4"
     },
     "require-dev": {
-        "brianium/paratest": "^6.3",
-        "facade/ignition": "^2.12",
-        "fakerphp/faker": "^1.15",
-        "friendsofphp/php-cs-fixer": "^3.1",
-        "itsgoingd/clockwork": "^5.1",
-        "mockery/mockery": "^1.4.2",
-        "nunomaduro/collision": "^5.9",
-        "phpunit/phpunit": "^9.5",
-        "wnx/laravel-stats": "^2.5.2"
+        "brianium/paratest": "^v6.6.1",
+        "facade/ignition": "^2.17.6",
+        "fakerphp/faker": "^v1.20.0",
+        "friendsofphp/php-cs-fixer": "^3.9.5",
+        "itsgoingd/clockwork": "^v5.1.6",
+        "mockery/mockery": "^1.5.0",
+        "nunomaduro/collision": "^v5.11.0",
+        "phpunit/phpunit": "^9.5.21",
+        "wnx/laravel-stats": "^v2.9.2"
     },
     "config": {
         "preferred-install": "dist",
