@@ -48,7 +48,7 @@ class RedList extends Model
     }
 
     /**
-     * Countries for reference local id's
+     * Countries for reference local id's.
      */
     public function countries()
     {
