@@ -20,14 +20,11 @@ return [
         'read' => 'Pročitaj vijest',
     ],
 
-    'field_observations_import' => [
+    'taxa_import' => [
         'short_info' => 'Ukoliko želite da uvezete podatke iz tablice, potrebno '.
-            'je da ona bude sačuvana kao CSV datoteka. Nakon izbora datoteke, treba '.
+            'je da ona bude sačuvana kao CSV/XLSX datoteka. Nakon izbora datoteke, treba '.
             'da uskladite redosled kolona u Biologeru tako da on odgovara redosledu u '.
-            'tablici i da odaberete koje kolone želite da uvezete. Spisak taksona mora '.
-            'da prati taksonomiju Biologer baze podataka, a spisak vrednosti za svaku '.
-            'kolonu (npr. stadijumi, pol, licenca) mora biti dat na osnovu vrednosti '.
-            'na engleskom jeziku.',
+            'tablici i da odaberete koje kolone želite da uvezete.',
     ],
 
     'taxa' => [
