@@ -20,14 +20,11 @@ return [
         'read' => 'Прочитај вест',
     ],
 
-    'field_observations_import' => [
+    'taxa_import' => [
         'short_info' => 'Уколико желите да увезете податке из таблице, потребно '.
-            'је да она буде сачувана као CSV датотека. Након избора датотеке, '.
+            'је да она буде сачувана као CSV/XLSX датотека. Након избора датотеке, '.
             'треба да ускладите редослед колона у Биологеру тако да он одговара '.
-            'редоследу у таблици и да одаберете које колоне желите да увезете. '.
-            'Списак таксона мора да прати таксономију Биологер базе података, а '.
-            'списак вредности за сваку колону (нпр. стадијуми, пол, лиценца) мора '.
-            'бити дат на основу вредности на енглеском језику.',
+            'редоследу у таблици и да одаберете које колоне желите да увезете. ',
     ],
 
     'taxa' => [
